requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-multipart==0.0.6
pillow==10.0.0
torch==2.0.1
transformers==4.32.1
langchain-google-genai==0.0.5
langchain-core==0.1.4
python-dotenv==1.0.0
